feat(runtime): move app_probes.go to internal/runtime (AS-151)
Migrates availability and Wave-2 enrichment probe logic from
internal/tui/app_probes.go to the platform-agnostic runtime package.

- internal/runtime/probes.go: Core methods (ProbeResourceAvailability,
  ProbeEnrichment, DemoPrefetchCounts, LoadAvailabilityCache,
  SaveAvailabilityCache, BuildEnrichQueue, BuildResourceCacheSnapshot).
  No Bubble Tea imports.
- internal/tui/probe_adapter.go: thin Bubble Tea adapter; wraps Core
  methods in tea.Cmd factories, converts results to TUI message types.
  Keeps refreshResourceListWithEnrichmentRerun (TUI-only concern).
- internal/tui/app.go: add core *runtime.Core field, wire via shared
  *session.Session in New().
- internal/tui/app_probes.go: deleted (content moved to above).
